refactor: replace direcotry
